move test_typing.py into test_base_config.py
